pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "dailyinfo"
version = "0.3.0"
description = "Daily briefing pipeline: scrape → AI summary → Discord push"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "BSD-3-Clause" }
authors = [{ name = "Wenyu Ouyang" }]
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "click>=8.1",
    "websocket-client>=1.9.0",
    "discord-py>=2.3",
    "httpx[socks]>=0.27",
    "ddgs>=9.14.1",
    "browser-use==0.12.6",
]

[project.optional-dependencies]
dev = [
    "black>=24",
    "ruff>=0.4",
    "pytest>=8",
]
paper = [
    "browser-use",
    "langchain-anthropic",
]

[project.scripts]
dailyinfo = "scripts.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*", "dailyinfo_fetcher*"]

[tool.uv]
dev-dependencies = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
